+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+## Added
+- Season 3 Curios
+- Shrieking Quartz (Nemesis lure)
+
 ## [1.10.3] 2025-08-06
 ## Fixed
 - Removed some deleted items